+++ Bonobo.Git.Server.Test/TestCategories.cs
@@ -8,7 +8,15 @@ namespace Bonobo.Git.Server.Test
 {
     public static class TestCategories
     {
-        public const string ClAndWebIntegrationTest = "ClAndWebIntegrationTest";
-        public const string WebIntegrationTest = "WebIntegrationTest";
+        public const string IntegrationTest = "IntegrationTest";
+
+        // User data storage type
+        public const string StorageInternal = "StorageInternal";
+        public const string StorageActiveDirectory = "StorageActiveDirectory";
+
+        // Authentication types
+        public const string AuthForms = "AuthForms";
+        public const string AuthWindows = "AuthWindowsAuth";
+        public const string AuthADFS = "AuthADFS";
     }
 }