+++ changelog.md
@@ -28,6 +28,7 @@ backup of your App_Data folder before you update.
 * The linkify option for commit messages can now be set on a per repo basis
 * Verifies that username, repository and team names are unique as you type them
 * Verifies as you type regex entered for linkification
+* Added icons to repository listing indicating anonymous push/pull status
 
 ### Bugfixes
 
@@ -36,6 +37,8 @@ backup of your App_Data folder before you update.
 * Correct error reported to Git clients for access to non-existent repo #447
 * Bonobo can start enough to allow access to settings if git directory is mis-configured #451
 * Bonobo doesn't run on systems with particular machine.config RoleManager settings #486
+* The repository logo was not visible in all views
+* Fixed inability to browse folders in branches other then master #541
 
 ### Code improvements